+++ src/headless/plugins/version/types.ts
@@ -1,0 +1,5 @@
+export type SoftwareVersion = {
+    name: string | null;
+    version: string | null;
+    os: string | null;
+};