+++ src/headless/types/plugins/version/types.d.ts
@@ -1,0 +1,6 @@
+export type SoftwareVersion = {
+    name: string | null;
+    version: string | null;
+    os: string | null;
+};
+//# sourceMappingURL=types.d.ts.map